requirements.txt
openai
python-dotenv
streamlit
requests
pandas
tenacity
pytrends
google-genai
streamlit
requests
python-dotenv
pytrends

